(PC-24469) move RecommendationAPIParams to /shared/offer types
